Add files via upload
The application has two activities. It displays personal information and another demonstrates activity callbacks.
Personal Information Activity: When you launch the application, it displays personal information. It contains fields for name, registration number and photo.
Callback Activity: The user can navigate to the second activity by clicking a button provided in the first activity. The second activity demonstrates four activity callbacks, such as onCreate, onStart, onResume, onPause, onStop, or onDestroy. Each callback represents a specific event in the lifecycle of an activity. When each callback is triggered, a Toast message is displayed on the screen to inform the user about the current state of the activity.
The application allows a user to view personal information in the first activity. It also demonstrates activity callbacks in the second activity. The user can navigate between the activities using buttons and the functionality of the application.